MODE-2012 Added test case that demonstrates clone problem
The problem with workspace clone is that references are being copied incorrectly. The resulting
references will keep pointing to the nodes in the source workspace, rather than the (newly cloned)
nodes in the target workspace.

This commit only adds a test case that fails with the problem; it does not yet fix the problem.