MODE-1467 Added ability to explicitly invoke sequencer within session
ModeShape now supports applications explicitly invoking sequencing
from within the scope of a session. There are several differences with
using sequencers in this style:

- A specific sequencer must be named, and only that sequencer will be
invoked.
- The path expressions of that configured sequencer are not used in
any way. All other configured properties are still used.
- The sequencer's accepted MIME types are still respected. If the
input content's MIME type is known and is not in the sequencer's set
of supported MIME types, the sequencing operation will return immediately.
- The caller must supply the property containing the value to be
sequenced.
- The caller must supply a new node into which the sequencer output
will be placed.
- The session used to sequence contains all sequencing output as
transient state; the caller is expected to call 'session.save()' if
the output is to be persisted.