MODE-2072: Changed the web explorer module to be two modules
This is required since the EAP-based explorer WAR file can be built
without the ModeShape libraries, since they are provided in the
subsystem's modules in the kit. However, all of our other WAR modules
under "web" directory all embed the ModeShape libraries. Thus,
this commit splits the "web/modeshape-web-explorer" into two modules:
the "web/modeshape-web-explorer-war" that does not embed the
ModeShape libraries (and which is used by the EAP explorer module)
and the "web/modeshape-web-explorer-war" that does embed the
ModeShape libraries (like the other "web" WAR modules).