MODE-2122 Added some retry logic in case we're reading the parent's ChildReference objects while it is being modified
Most of the time, this shouldn't be a problem. But the changes make the code tolerant of such rare concurrent modifications
by simply retrying. There is almost no extra cost to doing this.

The stress test added in the previous commit used to periodically demonstrate the concurrent modification exception.
After the changes, the exception no longer occurs.