MODE-2151 Added support for CHILDCOUNT dynamic operand
Pretty basic support that should prove quite useful in certain situations. This may be relatively
expensive when the repository has nodes with lots of children since it requires loading the parent
node's child references in order to obtain the count. The CHILDCOUNT criteria would therefore work
much better/faster as filtering criteria in a query that already defines criteria that indexes can
use.