MODE-2285 More correctios to the index provider behavior in the Wildfly subsystem.
After these changes, any index provider or index definitions in the subsystem XML remain there and unmodified even after shutdown.
The problem was the 'relative-to' and 'path' index provider properties were not in the list of model attributes, and thus ignored
(in some scenarios only).

Also fixed a potential NPE in the IndexProvider base class.