MODE-2309 Corrected reading and writing JSON files when escaped characters are used.
This fixes the backup and restore issue. Note that ModeShape uses BSON when storing nodes in Infinispan, so this
should not affect persistent storage in a cache store. ModeShape uses JSON for the file connector (extra properties)
and for the REST service; neither of these should be impacted negatively by this change, though like the backup
and restore this change should fix any as-of-yet unseen issues related to escaped characters.

Note that it is not possible to just update the Restore functionality, so with this change new backups will have
to be created. This change does need to be applied to the 3.x branch as well.